requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-dotenv==1.0.1
Werkzeug==3.0.2
zipp==3.18.1
opencv-python==4.9.0.80
ultralyticsplus==0.1.0
numpy==1.26.4
psutil==5.9.8
python-dotenv==1.0.1
